wordpress-automation/requirements.txt
# Core framework
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.3.0

# AI APIs
google-generativeai>=0.7.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# Data models
pydantic>=2.5.0

# Environment management
python-dotenv>=1.0.0

# Date/time handling
python-dateutil>=2.8.0

# Text processing
beautifulsoup4>=4.12.0
markdown>=3.5.0

# Logging
loguru>=0.7.0

# Image processing (for future image nodes)
pillow>=10.0.0
requests>=2.31.0